requirements.txt
numpy==1.26.4
torch==2.2.2
sentence-transformers==2.7.0
transformers==4.41.1
uvicorn==0.29.0
fastapi==0.111.0
safetensors==0.4.3
scipy==1.13.1
pydantic==2.7.1
typing-extensions>=4.5.0

python-dotenv==1.0.1
PyPDF2==3.0.1
requests==2.31.0
psycopg2-binary>=2.9.6,<2.9.10
pgvector==0.2.3
groq==0.5.0
httpx==0.27.0